daemon: Switch all engines to be plugins
