+++ daemon/engines/deezer_engine.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from collections.abc import Callable
+from engine import Engine, SongChoice
 
 import re
 import deemix
@@ -7,66 +7,16 @@ import deemix.settings
 from deemix.downloader import Downloader
 from deezer import Deezer, TrackFormats
 from deezer.errors import DataException
-#from youtube_dl import YoutubeDL
-#from youtube_search import YoutubeSearch
 
-from song import Song
-from library import Library
-from playlist import Playlist
-
-@dataclass
-class SongChoice:
-    title: str
-    artist: str
-    # Function that calls the callback with the filename when done
-    # (usually after downloading but not necessarily)
-    def choose(self, callback):
-        raise Exception("Choose was called but not implemented")
-
-    def __str__(self):
-        return f"**{self.title}** by *{self.artist}*"
-
-class Player:
-    name: str = None
-    description: str = None
-    command: str = None
-
-    def search(self, query: str):
-        pass
-
-class LibraryPlayer(Player):
-    name = "Library"
-    description = "Plays from the downloaded songs"
-    command = "lib"
-
-    def __init__(self, library):
-        self.library = library
-
-    @dataclass
-    class LibrarySongChoice(SongChoice):
-        song: Song
-
-        def choose(self):
-            return self.song.path
-
-    def search(self, query: str):
-        return [
-            self.LibrarySongChoice(
-                title = s.title,
-                artist = s.artist,
-                song = s,
-            )
-            for s in self.library.search(query)
-        ]
-
-class DeezPlayer(Player):
+class DeezerEngine(Engine):
     name = "Deezer"
     description = "Plays from deezer"
-    command = "deez"
 
-    def __init__(self, arl):
+    def __init__(self, library, options):
+        if not 'arl' in options:
+            raise Exception('No arl given to deezer engine, but required to function')
         self.dz = Deezer()
-        self.dz.login_via_arl(arl)
+        self.dz.login_via_arl(options['arl'])
 
     @dataclass
     class DeezSongChoice(SongChoice):
@@ -128,21 +78,3 @@ class DeezPlayer(Player):
             link = result['link']
         )]
 
-#class YoutubePlayer(Player):
-#    name = "Youtube"
-#    description = "Plays from youtube"
-#    command = "yt"
-#
-#    def __init__(self):
-#        self.ytdl = YoutubeDL()
-#
-#    def search(self, query: str):
-#        results = YoutubeSearch(query).videos
-#
-#        def download(link: str):
-#            pass
-#
-#        return [
-#            SongChoice(x['title'], x['channel'], lambda: download(x['id']))
-#            for x in results[0:9]
-#        ]
